Add function on_most_keys()
A utility function that is like keyboard_check_pressed(vk_anykey)
but ignores certain keys commonly used during development.
In the case of Shift and Control, the main annoyance is that they were
getting picked up while I was modifying code on the fly via GMLive.
In the case of Alt and PrtScr, I usually use Alt+PrtScr when I want to
take screenshots.